feat: separate i18n common
