+++ app/i18n/locales/en/common/days.ts
@@ -1,0 +1,9 @@
+export const days = {
+    monday: "Mon",
+    tuesday: "Tue",
+    wednesday: "Wed",
+    thursday: "Thu",
+    friday: "Fri",
+    saturday: "Sat",
+    sunday: "Sun",
+};
