+++ app/i18n/locales/en/common/general.ts
@@ -1,0 +1,22 @@
+export const general = {
+    class: "Class",
+    professor: "Professor",
+    description: "Description",
+    grade: "Grade",
+    load: "Load",
+    speech: "Speech",
+    upload: "Upload",
+    numClasses: "Lec. Hours",
+    numLabs: "Lab Hours",
+    credit: "Credits",
+    cancel: "Cancel",
+    save: "Save",
+    edit: "Edit",
+    hours: "Hours",
+    minutes: "Min",
+    seconds: "Sec",
+    am: "AM",
+    pm: "PM",
+    year: "year",
+    language: "Language",
+};
