+++ app/i18n/locales/en/common/professors.ts
@@ -1,0 +1,4 @@
+export const professors = {
+    over: " and ",
+    people: " others",
+};
