+++ app/i18n/locales/en/common/review.ts
@@ -1,0 +1,4 @@
+export const review = {
+    writingPlaceholder: "Please describe your evaluation of grades, load, etc.",
+    like: "Like ",
+};
