+++ app/i18n/locales/en/common/search.ts
@@ -1,0 +1,11 @@
+export const search = {
+    search: "Search",
+    cancel: "Cancel",
+    level: "Grade",
+    department: "Department",
+    term: "Term",
+    all: "All",
+    type: "Groups",
+    placeholder: "Enter keywords",
+    empty: "Please select at least one search filter",
+};
