+++ app/i18n/locales/en/common/semesters.ts
@@ -1,0 +1,6 @@
+export const semesters = {
+    spring: "Spring",
+    summer: "Summer",
+    fall: "Fall",
+    winter: "Winter",
+};
