+++ app/i18n/locales/en/common/term.ts
@@ -1,0 +1,5 @@
+export const term = {
+    "3yearsShort": "In-3-Years",
+    "1yearShort": "In-1-Year",
+    thisSemesterShort: "This-Semester",
+};
