+++ app/i18n/locales/ko/common/days.ts
@@ -1,0 +1,9 @@
+export const days = {
+    monday: "월요일",
+    tuesday: "화요일",
+    wednesday: "수요일",
+    thursday: "목요일",
+    friday: "금요일",
+    saturday: "토요일",
+    sunday: "일요일",
+};
