+++ app/i18n/locales/ko/common/general.ts
@@ -1,0 +1,22 @@
+export const general = {
+    class: "분류",
+    professor: "교수",
+    description: "설명",
+    grade: "성적",
+    load: "널널",
+    speech: "강의",
+    upload: "업로드",
+    numClasses: "강의시간",
+    numLabs: "실험",
+    credit: "학점",
+    cancel: "취소",
+    save: "저장",
+    edit: "수정",
+    hours: "시간",
+    minutes: "분",
+    seconds: "초",
+    am: "오전",
+    pm: "오후",
+    year: "연도",
+    language: "언어",
+};
