+++ app/i18n/locales/ko/common/professors.ts
@@ -1,0 +1,4 @@
+export const professors = {
+    over: " 외 ",
+    people: "명",
+};
