+++ app/i18n/locales/ko/common/review.ts
@@ -1,0 +1,4 @@
+export const review = {
+    writingPlaceholder: "학점, 로드 등의 평가에 대하여 설명해주세요.",
+    like: "좋아요",
+};
