+++ app/i18n/locales/ko/common/search.ts
@@ -1,0 +1,11 @@
+export const search = {
+    search: "검색",
+    cancel: "취소",
+    level: "학년",
+    department: "학과",
+    term: "기간",
+    all: "전체",
+    type: "분류",
+    placeholder: "과목명, 교수명을 등을 검색해보세요",
+    empty: "검색 조건을 하나 이상 선택해주세요",
+};
